projet_ATOS/cloud_function/requirements.txt
google-cloud-storage==2.4.0
google-cloud-pubsub==2.13.0 
apache-beam==2.40.0
schema==0.7.5
google-apitools==0.5.32
oauth2client==4.1.3
google-api-python-client==2.52.0